requirements.txt
BaseCon
Decibel
KVS
Msgr
mysql-connector-python
pymemcache
redis
SeCo
